+++ CHANGELOG.md
@@ -10,6 +10,10 @@
 
   The second (`scoped_metric`) and third (`elapsed`) arguments in the `NewRelic::Agent::Datastores.notice_sql` method are deprecated. They have not been used by the method for some time. Instead, these values will be set based on the current segment when the API is called. [PR#3345](https://github.com/newrelic/newrelic-ruby-agent/pull/3345)
 
+- **Feature: Deprecation notice for second argument in Datastores.notice_statement API**
+
+  The second (`elapsed`) argument in the `NewRelic::Agent::Datastores.notice_statement` method is deprecated. It has not been used by the method for some time. Instead, this value will be set based on the current segment when the API is called. [PR#3346](https://github.com/newrelic/newrelic-ruby-agent/pull/3346)
+
 ## v9.23.0
 
 - **Feature: Add sidekiq.ignore_retry_errors configuration option**